Create relocation records of type TR_RecompQueuedFlag
In OpenJ9 the JProfiling framework is used to create profiled code
which contains references to internal data structures like block
frequencies stored in `TR_BlockFrequencInfo` class, or the flag for
checking if the method is already in recompilation queue.
To support AOT compilation for such code, we need to generate relocation
records for such data structures referenced by the compiled code.
OpenJ9 PR https://github.com/eclipse/omr/pull/5202 creates the necessary
structures representing the relocation record for this data.
This commit adds the code for generating the relocation records
specifically for recompilation queue flag.

Signed-off-by: Ashutosh Mehra <mehra.ashutosh@ibm.com>